Added Phase1 and SYNC
Ended up digging into the 6502 documentation to understand why SYNC and Phase1 are important. Although the IIc and IIe do not use these signals, they do have a function. Sync tells us when the cpu is executing an action, on the system. Phase1 is the inverse of Phase0 or Phase2, and is the time when the Apple II series performs memory actions. Video phase addressing traffic is not visible to the CPU in the Apple II series, the Option ports need to be used to grab those signals. RAn